SearchForm: restructure as dir
