+++ src/screens/PeopleSearchScreen.js
@@ -27,7 +27,7 @@ import { ScrollView, FlatList } from 'react-native-gesture-handler';
 import PersonRow from '../components/Person/PersonRow';
 import headerConfig from '../helpers/headerConfig';
 import constants from '../helpers/constants';
-import SearchForm from '../components/SearchForm/SearchForm';
+import SearchForm from '../components/SearchForm/index.jsx';
 import { sendEvent } from '../helpers/createEvent';
 import Loader from '../components/Loader/index.jsx';
